+++ Makefile
@@ -7,17 +7,25 @@ BUILD_DIR = build
 SRC_DIR = src
 
 # 目标文件
-TARGET = $(BUILD_DIR)/main.o
+TARGET1 = $(BUILD_DIR)/smtp.o
+TARGET2 = $(BUILD_DIR)/pop3.o
 
-# 源文件
-SRCS = $(wildcard $(SRC_DIR)/*.c)
+# smtp源文件
+SRCS = $(wildcard $(SRC_DIR)/*.c) smtp_main.c
+
+# pop3源文件
+POP_SRCS = $(wildcard $(SRC_DIR)/*.c) pop3_main.c
 
 the_main:
-	$(CC) $(CFLAGS) $(SRCS) -o $(TARGET)
+	$(CC) $(CFLAGS) $(SRCS) -o $(TARGET1)
+	$(CC) $(CFLAGS) $(POP_SRCS) -o $(TARGET2)
 
-run: the_main
-	$(TARGET)
+smtp: the_main
+	$(TARGET1)
 
-test:
+pop: the_main
+	$(TARGET2)
+
+debug:
 	$(CC) $(CFLAGS) $(SRCS) -g
 	$(DB) ./a.out
