+++ database/zhangsan_14.mail
@@ -1,0 +1,12 @@
+Message-ID: <001101c93cc4$1ed2ae30$5400a8c0@liuxiaoforever>\r\n
+From: "zhangsan" <zhangsan@qq.com>\r\n
+To: 123<123@qq.com>\r\n
+Subject:123 0 0 Control Success! \r\n
+Date: Wed Apr  3 15:52:42 2024\r\n
+MIME-Version: 1.0\r\n
+Content-Type: multipart/alternative; \r\n
+X-Priority: 3..X-MSMail-Priority: Normal\r\n
+X-Mailer: Microsoft Outlook Express 6.00.2900.3138\r\n
+X-MimeOLE: Produced By Microsoft MimeOLE V6.00.2900.3198\r\n
+\r\n
+This is response of 123@qq.com's device!\r\n
