--- src/pop3_main.c
@@ -1,1 +1,0 @@
-#include "common.h"
